Restructure project paths
Moved all apps to the apps folder and created the packages folder for things that will be shared.

Added tsconfig and package.json to all the apps and projects. Added Eslint to the project.

Moved django backend to apps/backend-legacy/ moved bots to apps/bot/

Now using pnpm instead on npm.

Added a basic nestjs files to apps/api/ and basic drizzle orm to packages/db/